Services/gRPC/S2VT_VideoCaptioning/requirements.txt
grpcio==1.12.1                
grpcio-tools==1.12.1
pytube==9.3.3
opencv-python==3.4.3.18
h5py==2.7.1
matplotlib==2.2.2
pathlib==1.0.1